+++ src/pages/doors/index.js
@@ -1,12 +1,12 @@
 import { useState, useEffect } from 'react';
 
-import { getDoorsCount, setDoors }       from '../../redux/slices/catalogSlice';
-import { useAppDispatch } from './../../redux/hook';
-import { useAppSelector } from '../../redux/hook';
-import Catalog            from './../../components/catalog/catalog';
-import DoorsHeader        from '../../components/doorsHeader/doorsHeader';
-import MainContainer      from '../../components/mainLayout/mainLayout';
-import Pagination from './../../components/shared/pagination/pagination';
+import { getDoorsCount, setDoors } from '../../redux/slices/catalogSlice';
+import { useAppDispatch }          from './../../redux/hook';
+import { useAppSelector }          from '../../redux/hook';
+import Catalog                     from './../../components/catalog/catalog';
+import DoorsHeader                 from '../../components/doorsHeader/doorsHeader';
+import MainContainer               from '../../components/mainLayout/mainLayout';
+import Pagination                  from './../../components/shared/pagination/pagination';
 
 export const getStaticProps = async () => {
     const basePath = process.env.NEXT_PUBLIC_API_LINK;
@@ -22,8 +22,7 @@ export const getStaticProps = async () => {
 const Doors = ({ serverDoors }) => {
     const dispatch = useAppDispatch();
     const [localDoors, setLocalDoors] = useState(serverDoors);
-    const storageDoors = useAppSelector(state => state.catalog.doors);
-    const currentPage = useAppSelector(state => state.catalog.currentPage);
+    const { currentPage, doors } = useAppSelector(state => state.catalog);
 
     useEffect(() => {
         dispatch(setDoors(localDoors));
@@ -31,8 +30,8 @@ const Doors = ({ serverDoors }) => {
     }, []);
 
     useEffect(() => {
-        setLocalDoors(storageDoors);
-    }, [storageDoors]);
+        setLocalDoors(doors);
+    }, [doors]);
 
     return (
         <MainContainer keywords="" title={`Каталог дверей страница ${currentPage}`}>